src/api/db/scripts/migrate.py
import os
from urllib.parse import quote_plus

from dotenv import load_dotenv
from yoyo import get_backend, read_migrations

load_dotenv()

POSTGRES_USER = os.environ.get("POSTGRES_USER") or "postgres"
POSTGRES_PASSWORD = os.environ.get("POSTGRES_PASSWORD") or "password"
POSTGRES_HOST = os.environ.get("POSTGRES_HOST") or "localhost"
POSTGRES_PORT = os.environ.get("POSTGRES_PORT") or "5432"
POSTGRES_DB = os.environ.get("POSTGRES_DB") or "postgres"


connect_string = f"postgresql://{POSTGRES_USER}:{quote_plus(POSTGRES_PASSWORD)}@{POSTGRES_HOST}:{POSTGRES_PORT}/{POSTGRES_DB}"

backend = get_backend(connect_string)

migrations = read_migrations("migrations")
with backend.lock():
    backend.apply_migrations(backend.to_apply(migrations))
