pyproject.toml
[tool.poetry]
name = "narrative-md"
version = "0.1.0"
description = ""
authors = ["Gregor Lämmel <GrgrLmml@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.35.0"
streamlit-webrtc = "^0.47.7"
watchdog = "^4.0.1"
jupyterlab = "^4.2.2"
pydub = "^0.25.1"
openai = "^1.34.0"
fastapi = "^0.111.0"
asyncpg = "^0.29.0"
gunicorn = "^22.0.0"
uvicorn = "^0.30.1"
yoyo-migrations = "^8.2.0"
psycopg2 = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
